Move _maybe_shard_state_dict_for_tp back
It is imported from save_and_load.py in the Transformers PEFT
integration and as such should stay there.
